Fix bug when editing habit and fix tests
